+++ pkg/backup/prometheus.go
@@ -1,4 +1,4 @@
-package scheduler
+package backup
 
 import (
 	"regexp"
