+++ pkg/backup/restic.go
@@ -1,4 +1,4 @@
-package scheduler
+package backup
 
 import (
 	"fmt"
