+++ pkg/cmds/root.go
@@ -46,7 +46,7 @@ func NewCmdStash(version string) *cobra.Command {
 
 	rootCmd.AddCommand(v.NewCmdVersion())
 	rootCmd.AddCommand(NewCmdRun(version))
-	rootCmd.AddCommand(NewCmdSchedule())
+	rootCmd.AddCommand(NewCmdBackup())
 	rootCmd.AddCommand(NewCmdRecover())
 	return rootCmd
 }
